Brandon, read this.
I have a plan for the power ups and spawning them and whatnot, so you
should work on something else. All powerups can be simulated with 1-7 as
seen on the PlayerVariables class. If you run out of things to do, go
for it, but my eyes won't stay open. Otherwise, focus on getting the
Oculus working with my new and improved shooting :tongue: , and UI
stuff. Otherwise, feel free to head wherever you want. I'm probably
going to be asleep for a bit.